refactor User, Group, Doc
