+++ richreview_core/node_server/db_schema.md
@@ -12,42 +12,36 @@ userid is a sha1 hash with salt from netid
 @member email    {string} required - email of user
 @member groupNs  {string|Array<string>} required - array of groupid user is in
 
-userid is also added to email_user_lookup as a hash field
-
-## UBCUser
-
-**UBCUser** ( `ubc_user:<userid>` )
-userid is a sha1 hash with salt from netid  
-
-@type hash / class
 @member password   {string} - made from irreversible sha1 hash with salt from netid
 @member auth_level {string} - is one of "student", "instructor", or "admin"; refers to security level in terms of access to functionality (delete, doc creation, etc); also affects routing.
 @member first_name {string} optional - the first name of user
 @member last_name  {string} optional - the last name of user
 @member sid        {string} optional - the student id of user
 
+userid is also added to email_user_lookup as a hash field
+
 ## Course
 
-**Course** ( `ubc_course:<course-dept>:<course-number>` )
+**Course** ( `course:<course-dept>:<course-number>` )
 @type hash / class
 @member course_is_active {boolean} - true if course is active, false otherwise
 @member name  {string} - name of the course; defaults to `<course-dept> <course-number>`
 
 ## Course users
 
-**course instructor** ( `ubc_course:<course-dept>:<course-number>:instructors` ) is of type set containing instructors of course
+**course instructor** ( `course:<course-dept>:<course-number>:instructors` ) is of type set containing instructors of course
 
-**course active students** ( `ubc_course:<course-dept>:<course-number>:students:active` ) is of type set containing active students of course
+**course active students** ( `course:<course-dept>:<course-number>:students:active` ) is of type set containing active students of course
 
-**course blocked students** ( `ubc_course:<course-dept>:<course-number>:students:blocked` ) is of type set containing blocked students of course
+**course blocked students** ( `course:<course-dept>:<course-number>:students:blocked` ) is of type set containing blocked students of course
 
 ## Assignment
 
-**Assignment** ( `ubc_asgmt:<userid>:<course-dept>:<course-number>:<title-slug>` )
+**Assignment** ( `asgmt:<userid>:<course-dept>:<course-number>:<title-slug>` )
 userid      is a sha1 hash with salt from netid  
 title-slug  is the slug of the name of the title  
-Search for assignments belonging to [userid] by `keys ubc_asgmt:[userid]:*`  
-Search for assignments belonging to [course-dept] and [course-number] by `keys ubc_asgmt:*:[course-dept]:[course-number]:*`  
+Search for assignments belonging to [userid] by `keys asgmt:[userid]:*`  
+Search for assignments belonging to [course-dept] and [course-number] by `keys asgmt:*:[course-dept]:[course-number]:*`  
 
 @type    hash / class 
 @member title     {string}required - defaults to `<email-hash>_<timestamp>` 
@@ -85,7 +79,7 @@ Search for assignments belonging to [course-dept] and [course-number] by `keys u
   *
   * is_active   {string|boolean} - can access ; instructor and admin should ALWAYS be active
   * courses    {string|Array<string>} - an array of strings for courses of form `course:<course-dept>:<course-number>`
-  * portfolios {string|Array<string>} - an array of strings for portfolios of form `ubc_portfolio:<userid>:<course-dept>:<course-number>`
+  * portfolios {string|Array<string>} - an array of strings for portfolios of form `portfolio:<userid>:<course-dept>:<course-number>`
   *
   * first_name {string} optional - the first name of user
   * last_name  {string} optional - the last name of user
@@ -101,21 +95,21 @@ Search for assignments belonging to [course-dept] and [course-number] by `keys u
   */
 
  /**
-  * Portfolio ( ubc_portfolio:<userid>:<course-dept>:<course-number> )
-  * Search for portfolios belonging to [userid] by `keys ubc_portfolio:[userid]:*`
-  * Search for portfolios belonging to [course-dept] and [course-number] by `keys ubc_portfolio:*:[course-dept]:[course-number]`
+  * Portfolio ( portfolio:<userid>:<course-dept>:<course-number> )
+  * Search for portfolios belonging to [userid] by `keys portfolio:[userid]:*`
+  * Search for portfolios belonging to [course-dept] and [course-number] by `keys portfolio:*:[course-dept]:[course-number]`
   *
-  * assignments {string|Array<string>} - array of string `ubc_asgmt:<userid>:<course-dept>:<course-number>:<title>`
+  * assignments {string|Array<string>} - array of string `asgmt:<userid>:<course-dept>:<course-number>:<title>`
   *
   * TODO: this may not be necessary
   */
 
  /**
-  * Assignment ( ubc_asgmt:<userid>:<course-dept>:<course-number>:<title-slug> )
+  * Assignment ( asgmt:<userid>:<course-dept>:<course-number>:<title-slug> )
   * userid      is a sha1 hash with salt from netid
   * title-slug  is the slug of the name of the title
-  * Search for assignments belonging to [userid] by `keys ubc_asgmt:[userid]:*`
-  * Search for assignments belonging to [course-dept] and [course-number] by `keys ubc_asgmt:*:[course-dept]:[course-number]:*`
+  * Search for assignments belonging to [userid] by `keys asgmt:[userid]:*`
+  * Search for assignments belonging to [course-dept] and [course-number] by `keys asgmt:*:[course-dept]:[course-number]:*`
   *
   * type    hash / class
   *
